backend/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
cloudinary
python-dotenv